+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Unreleased
 - Renamed from Attributes.js to DDL.js.
+- Returns an instance of [`Ddl`][] from [`Ddl.sqlite3`][] and
+  [`Ddl.postgresql`][] for future compatibility.
+
+[`Ddl`]: https://github.com/moll/js-ddl/blob/master/doc/API.md#Ddl
+[`Ddl.postgresql`]: https://github.com/moll/js-ddl/blob/master/doc/API.md#Ddl.postgresql
+[`Ddl.sqlite3`]: https://github.com/moll/js-ddl/blob/master/doc/API.md#Ddl.sqlite3
 
 ## 0.1.337 (Oct 4, 2013)
 - First release.